% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_third_side.R
\name{calc_third_side}
\alias{calc_third_side}
\title{Calculate the Third Side of a Right Triangle}
\usage{
calc_third_side(a, b, c)
}
\arguments{
\item{a}{Numeric. One of the sides of the right triangle. Set to \code{NA} if this is the side to be calculated.}

\item{b}{Numeric. Another side of the right triangle. Set to \code{NA} if this is the side to be calculated.}

\item{c}{Numeric. The hypotenuse of the right triangle. Set to \code{NA} if this is the side to be calculated.}
}
\value{
The length of the missing side (\code{a}, \code{b}, or \code{c}) of the right triangle.
}
\description{
This function calculates the length of the third side of a right triangle when two sides are provided.
Given two sides (either \code{a} and \code{b} for the legs, or one leg and the hypotenuse \code{c}), the function will compute the missing side.
}
\examples{
# Calculate the hypotenuse when the two legs are known
calc_third_side(a = 3, b = 4, c = NA)  # Returns 5

# Calculate one leg when the other leg and the hypotenuse are known
calc_third_side(a = NA, b = 4, c = 5)  # Returns 3

# You must specify two sides for the calculation
calc_third_side(a = 3, b = NA, c = NA)  # Error, as only one side is provided


}
\keyword{Pythagoras}
\keyword{geometry,}
\keyword{right}
\keyword{triangle,}
