% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimmed_mean.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{Calculate Trimmed Mean}
\usage{
trimmed_mean(x, s, l)
}
\arguments{
\item{x}{A numeric vector from which to calculate the trimmed mean.}

\item{s}{An integer specifying the number of smallest values to trim.}

\item{l}{An integer specifying the number of largest values to trim.}
}
\value{
The trimmed mean of the vector \code{x} after trimming \code{s} smallest and \code{l} largest values.
}
\description{
This function calculates the trimmed mean of a numeric vector by removing a specified number of smallest and largest values.
}
\examples{
trimmed_mean(c(1, 2, 3, 4, 5, 6), s = 1, l = 1)  # Returns the mean of 2, 3, 4, 5
trimmed_mean(c(1,2), s=1, l=2)
trimmed_mean(c(1,2,3,4,5), s=1, l=2)
trimmed_mean(c(5,2,4,5,6,7), s=1, l=2)
}
